# docs/swagger.yaml
swagger: "2.0"
info:
  title: "API Raízes do Nordeste"
  description: "Sistema Multicanal de Gestão de Pedidos — canais APP, TOTEM, BALCAO, PICKUP e WEB"
  version: "1.0.0"
basePath: "/"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "Cole aqui: Bearer SEU_TOKEN"
...
paths:
...
  /pedidos:
    post:
      tags: ["Pedidos"]
      summary: "Criar pedido — canalPedido OBRIGATÓRIO"
      security:
        - Bearer: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            required: ["unidade_id", "canal_pedido", "itens"]
            properties:
              unidade_id:
                type: "integer"
                example: 1
              canal_pedido:
                type: "string"
                enum: ["APP", "TOTEM", "BALCAO", "PICKUP", "WEB"]
                example: "APP"
              itens:
                type: "array"
                items:
                  type: "object"
                  required: ["produto_id", "quantidade"]
                  properties:
                    produto_id:
                      type: "integer"
                      example: 1
                    quantidade:
                      type: "integer"
                      example: 2
              observacao:
                type: "string"
                example: "Sem pimenta"
      responses:
        201:
          description: "Pedido criado com sucesso"
          schema:
            type: "object"
            properties:
              id:
                type: "integer"
                example: 1
              canal_pedido:
                type: "string"
                example: "APP"
              status:
                type: "string"
                example: "AGUARDANDO_PAGAMENTO"
              valor_total:
                type: "number"
                example: 59.80
        401:
          description: "Não autenticado"
        404:
          description: "Produto ou unidade não encontrado"
        409:
          description: "Estoque insuficiente"
        422:
          description: "canalPedido ausente ou inválido"

    get:
      tags: ["Pedidos"]
      summary: "Listar pedidos com filtros"
      security:
        - Bearer: []
      parameters:
        - name: "canal_pedido"
          in: "query"
          type: "string"
          enum: ["APP", "TOTEM", "BALCAO", "PICKUP", "WEB"]
          example: "APP"
        - name: "status"
          in: "query"
          type: "string"
          enum: ["AGUARDANDO_PAGAMENTO", "PAGO", "EM_PREPARO", "PRONTO", "ENTREGUE", "CANCELADO"]
          example: "AGUARDANDO_PAGAMENTO"
        - name: "page"
          in: "query"
          type: "integer"
          default: 1
        - name: "limit"
          in: "query"
          type: "integer"
          default: 10
      responses:
        200:
          description: "Lista de pedidos"
        401:
          description: "Não autenticado"
  /pedidos/{id}:
    get:
      tags: ["Pedidos"]
      summary: "Detalhe de um pedido"
      security:
        - Bearer: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
          example: 1
      responses:
        200:
          description: "Pedido encontrado"
        401:
          description: "Não autenticado"
        403:
          description: "Sem permissão"
        404:
          description: "Pedido não encontrado"
  /pedidos/{id}/status:
    patch:
      tags: ["Pedidos"]
      summary: "Atualizar status do pedido (GERENTE/COZINHA/ADMIN/ATENDENTE)"
      security:
        - Bearer: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
          example: 1
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            required: ["status"]
            properties:
              status:
                type: "string"
                enum: ["EM_PREPARO", "PRONTO", "ENTREGUE", "CANCELADO"]
                example: "EM_PREPARO"
      responses:
        200:
          description: "Status atualizado"
          schema:
            type: "object"
            properties:
              id:
                type: "integer"
                example: 1
              status:
                type: "string"
                example: "EM_PREPARO"
        401:
          description: "Não autenticado"
        403:
          description: "Sem permissão"
        404:
          description: "Pedido não encontrado"
        409:
          description: "Transição de status inválida"
  /pedidos/{id}/cancelar:
    post:
      tags: ["Pedidos"]
      summary: "Cancelar pedido (operação sensível, exige motivo)"
      description: "Cancela um pedido nos status AGUARDANDO_PAGAMENTO ou PAGO. Devolve o estoque para a unidade e registra log de auditoria com categoria CANCELAMENTO. Cliente só pode cancelar os próprios pedidos."
      security:
        - Bearer: []
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
          example: 1
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            required: ["motivo"]
            properties:
              motivo:
                type: "string"
                description: "Justificativa do cancelamento (mínimo 5 caracteres)"
                example: "Cliente desistiu antes do preparo"
      responses:
        200:
          description: "Pedido cancelado com sucesso (estoque devolvido)"
          schema:
            type: "object"
            properties:
              mensagem:
                type: "string"
                example: "Pedido cancelado com sucesso."
              motivo:
                type: "string"
                example: "Cliente desistiu antes do preparo"
              estoque_devolvido:
                type: "boolean"
                example: true
        401:
          description: "Não autenticado"
        403:
          description: "Cliente tentando cancelar pedido de outro usuário"
        404:
          description: "Pedido não encontrado"
        409:
          description: "Pedido não pode ser cancelado no status atual (EM_PREPARO ou superior)"
        422:
          description: "Motivo ausente ou muito curto"
...